Move streams to /services/streams folder
