+++ app/services/index.js
@@ -35,11 +35,11 @@ const SendTransactionFileService = require('./send_transaction_file.service')
 const ShowAuthorisedSystemService = require('./show_authorised_system.service')
 const ShowRegimeService = require('./show_regime.service')
 const ShowTransactionService = require('./show_transaction.service')
-const StreamReadableDataService = require('./stream_readable_data.service')
-const StreamReadableRecordsService = require('./stream_readable_records.service')
-const StreamTransformCSVService = require('./stream_transform_csv.service')
-const StreamTransformUsingPresenterService = require('./stream_transform_using_presenter.service')
-const StreamWritableFileService = require('./stream_writable_file.service')
+const StreamReadableDataService = require('./streams/stream_readable_data.service')
+const StreamReadableRecordsService = require('./streams/stream_readable_records.service')
+const StreamTransformCSVService = require('./streams/stream_transform_csv.service')
+const StreamTransformUsingPresenterService = require('./streams/stream_transform_using_presenter.service')
+const StreamWritableFileService = require('./streams/stream_writable_file.service')
 const TransformRecordsToFileService = require('./transform_records_to_file.service')
 const ViewBillRunService = require('./view_bill_run.service')
 const ViewBillRunInvoiceService = require('./view_bill_run_invoice.service')
